pyproject.toml
[tool.poetry]
name = "metaphor-connectors"
version = "0.2.20"
description = ""
authors = ["Metaphor <dev@metaphor.io>"]
packages = [
    { include = "metaphor" },
]

[tool.poetry.dependencies]
asyncpg = { version = "~=0.22.0", optional = true }
aws-assume-role-lib = "^1.3.0"
boto3 = "~=1.17.53"
botocore = "~=1.20.53"
canonicaljson = "^1.4.0"
dataclasses-json = "^0.5.2"
fastjsonschema = "~=2.15.1"
google-api-python-client = { version = "~=2.2.0", optional = true }
google-auth-oauthlib = { version = "~=0.4.4", optional = true }
google-cloud-bigquery = { version = "^2.21.0", optional = true }
lkml = { version = "^1.1.0", optional = true }
looker-sdk = { version = "^21.4.1", optional = true }
metaphor-models = "^0.0.4"
python = ">=3.7,<3.10"
python-dateutil = "~=2.8.1"
requests = "^2.25.1"
slack-sdk = { version = "^3.5.1", optional = true }
smart-open = "^5.0.0"
snowflake-connector-python = { version = "~=2.4.1", optional = true }
sql-metadata = { version = "^2.0.0", optional = true }

[tool.poetry.extras]
all = [
  "asyncpg",
  "google-api-python-client",
  "google-auth-oauthlib",
  "google-cloud-bigquery",
  "lkml",
  "looker-sdk",
  "slack-sdk",
  "snowflake-connector-python",
  "sql-metadata",
]
bigquery = ["google-cloud-bigquery"]
dbt = []
postgresql = ["asyncpg"]
snowflake = ["snowflake-connector-python"]
google = ["google-api-python-client","google-auth-oauthlib"]
looker = ["lkml", "looker-sdk", "sql-metadata"]
slack = ["slack-sdk"]

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flakehell = "^0.9.0"
flake8 = "3.9.0"
freezegun = "^1.1.0"
isort = "^5.8.0"
mypy = "^0.910"
pytest = "^6.2.2"
pytest-asyncio = "^0.14.0"
types-freezegun = "^0.1.4"
types-requests = "^2.25.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
extend_skip = [
  ".serverless",
]

[tool.flakehell]
format = "colored"

[tool.flakehell.plugins]
pycodestyle = ["+*", "-E203", "-E501", "-W503"]
pyflakes = ["+*"]

[tool.mypy]
ignore_missing_imports = true
